Fixing a bunch of small bugs
Do RLE writing on demand so we can safeguard against the case where RLE
expands the input bytes.

Improve RLE a bit, use the full byte range when dealing with large
amounts of repeats.

Add tests for block writing to ensure overflow writes are handled
properly with the RLE step.