Fixed incorrect BWT results for a bunch of values.
The sorting was being done incorrectly when the lexigraphic comparison
terminated for values of different lengths. This just makes sure
comparisons are done for values of the same length. This does slow
things down by a ton, but that doesn't for now since the sorting backend
will be replaced anyway.