Moved code to internal packages to clean tree up
I did this before I rewrote the Huffman coding code so it works as
expected because it adds a bunch of types and routines which would just
clutter the unexported space. There was also just no reason for things
like crc and the bit writer to where it was.

A few other things changed but they're small.