Harden initial RLE.
Previously the initial RLE had a bunch of problems, mainly if a run
spanned across writes the encoder wouldn't account for that and the
result could be invalid. So instead of actually encoding the data and
dealing with the complexities of manipulation of the encoded bytes it
instead keeps a list of runs and then increments the previous runs
length if it spanned across writes. Also if overwrites occur it's now
more complex and actually counts the trimmed bytes as it goes rather
than just reslicing which could be tricky to get right. It still takes
into account of the possilibity that the resulting trimming could leave
a run with 4 bytes long which is invalid because that would make the
resulting encoding 5 bytes long, so it still ensures 3 byte runs are
used in that case to guard against expansion.

Also moved the test helper routines to their own internal package.